Update
- Updated confirmation window to contain optional dont show again checkbox
- Added LevelFile loading and made it threaded so the game loads instantly and doesn't wait
- WIP Continued implementing Rayman PC DOS
- Fixed no button onclick listeners not getting cleared
- Fixed edit not loading binary mappings due to load order
- Added preview animations to animator animations too
- Split tooltip frontend and settings to separate files
- Updated export binary window to have pixel scale be an advanced setting
- Added palette data sprites for WIP palette pick when loading binary animations
- Added moskito and ds rayman spritesets for mapping

+++ Assets/Scripts/Animation/Mapping.cs
@@ -111,7 +111,7 @@ public class Mapping
         if (targetImages.Count == 0)
         {
             Rayman1MSDOS.DesignObjects currObject = (Rayman1MSDOS.DesignObjects)Enum.Parse(typeof(Rayman1MSDOS.DesignObjects), _targetSpriteset);
-            targetImages = Rayman1BinaryAnimation.Instance.LoadSpritesetFromBinary((int)currObject);
+            targetImages = Rayman1BinaryAnimation.Instance.LoadSpritesetFromBinary(currObject);
         }
 
         for (int i = 0; i < MappingValues.Length; i++)
